+++ R/get_pvol_cz.R
@@ -55,13 +55,6 @@ get_pvol_cz <- function(radar, time, ..., call = rlang::caller_env()) {
   polar_volumes_tibble <- files_to_get |>
     dplyr::mutate(
       tempfile = purrr::map_chr(resp, purrr::chuck, "body"),
-      # add h5 how group as it seems to be missing
-      mut = purrr::map(tempfile, ~ {
-        hdf_connection <- rhdf5::H5Fopen(.x)
-        group <- rhdf5::H5Gcreate(hdf_connection, "how")
-        rhdf5::H5Fclose(hdf_connection)
-        rhdf5::H5Gclose(group)
-      }),
       pvol = purrr::map(tempfile, ~ bioRad::read_pvolfile(.x)),
       remove = purrr::map(tempfile, ~ file.remove(.x))
     )