+++ NEWS.md
@@ -1,3 +1,5 @@
+# getRad (development version)
+
 # getRad 0.2.0
 
 * New function `get_weather_radars()` retrieves metadata for OPERA weather radars (#15, #54).