+++ NEWS.md
@@ -1,5 +1,7 @@
 # getRad (development version)
 
+* How attribute is now present in Czech data (#102).
+
 # getRad 0.2.0
 
 * New function `get_weather_radars()` retrieves metadata for OPERA weather radars (#15, #54).