+++ NEWS.md
@@ -1,5 +1,7 @@
 # getRad (development version)
 
+* A bug (#101) in `get_vpts()` was fixed that caused the function to only return the first day of an interval, regardless of the length of the interval. (#105)
+
 # getRad 0.2.0
 
 * New function `get_weather_radars()` retrieves metadata for OPERA weather radars (#15, #54).