+++ NEWS.md
@@ -1,4 +1,4 @@
-# getRad (development version)
+# getRad 0.2.1
 
 * A bug (#101) in `get_vpts()` was fixed that caused the function to only return the first day of an interval, regardless of the length of the interval (#105).
 * Support downloading Swedish polar volume data (#96).